tests(plugins) remove duplication in udp server tests
Removes all repeated instances of customized UDP servers that only
differ in the number of packets received, making their behavior
more consistent in the process.

This modifies tests for `statsd` and `datadog` in `spec/`
(based on the changes to `spec-old-api/` in the `next` branch
at 51ea0847ef53cd525001e355b6283fd5eee75cfe).